+++ bbconf/models/bed_models.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import List, Optional, Union
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import BaseModel, ConfigDict, Field, field_validator
 
 from bbconf.const import DEFAULT_LICENSE
 
@@ -36,10 +36,14 @@ class BedClassification(BaseModel):
     name: Optional[str] = None
     genome_alias: str = None
     genome_digest: Union[str, None] = None
-    bed_type: str = Field(
-        default="bed3", pattern="^bed(?:[3-9]|1[0-5])(?:\+|$)[0-9]?+$"
+    bed_compliance: str = Field(
+        default="bed3", pattern=r"^bed(?:[3-9]|1[0-5])(?:\+|$)[0-9]?+$"
     )
-    bed_format: str = None
+    data_format: Union[str, None] = None
+    compliant_columns: int = 3
+    non_compliant_columns: int = 0
+
+    header: Union[str, None] = None  # Header of the bed file (if any)
 
     model_config = ConfigDict(extra="ignore")
 
@@ -134,17 +138,29 @@ class StandardMeta(BaseModel):
         "", description="Treatment of the sample (e.g. drug treatment)"
     )
 
-    global_sample_id: str = Field(
+    global_sample_id: Union[List[str], None] = Field(
         "", description="Global sample identifier. e.g. GSM000"
     )  # excluded in training
-    global_experiment_id: str = Field(
+    global_experiment_id: Union[List[str], None] = Field(
         "", description="Global experiment identifier. e.g. GSE000"
     )  # excluded in training
+
+    original_file_name: str = Field("", description="Original file name")
 
     model_config = ConfigDict(
         populate_by_name=True,
         extra="ignore",
     )
+
+    @field_validator("global_sample_id", "global_experiment_id", mode="before")
+    def ensure_list(cls, v: Union[str, List[str]]) -> List[str]:
+        if isinstance(v, str):
+            return [v]
+        elif isinstance(v, list):
+            return v
+        elif isinstance(v, type(None)):
+            return []
+        raise ValueError("values must be a string or a list of strings")
 
 
 class BedPEPHubRestrict(BedPEPHub):